requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
protobuf==3.20.1
PyMySQL==1.0.2
Werkzeug==2.2.2
